Move ServiceRegistration into its own file
- Separate service registration, so we can begin to move non-service
  related info back to the app config.
- newServiceRegistartion doesn't need to be a method of *Store.
